requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.3
openai>=1.0.0
twilio>=9.0.0
APScheduler>=3.10.0
python-dotenv>=1.0.0
apify-client>=1.6.0
